Create ThirdLargeNumArr[array].java
Third largest element
Given an array of distinct elements. Find the third largest element in it. 

Input:
The first line of input contains an integer T denoting the number of test cases. Then T test cases follow. The first line of each test case is N, size of the array. The second line of each test case contains N space separated values of the array a[].

Output:
Print the third largest element of the array. If the array has less than 3 elements print -1. 

Your Task:
You don't need to read input or print anything. Your task is to complete the function thirdLargest() which takes the array a[] and the size of the array as inputs and returns the third largest element in the array. Return -1 if there are less than 3 elements in the input.

Expected Time Complexity: O(N).
Expected Auxiliary Space: O(1).