Create InsertNodeSortedLL[LL].java
Insert in a Sorted List
Given a sorted singly linked list and a data, your task is to insert the data in the linked list in a sorted way i.e. order of the list doesn't change.

Input:
The function takes 2 arguments as input, reference pointer to the head of the sorted single linked list and an integer data value which is to be inserted in the list.
There are multiple test cases and for each test case, the function will be called separately.

Output:
For each test, the output will be space-separated integers denoting the values of the linked list.

User Task:
The task is to complete the function sortedInsert() which should insert the element in sorted Linked List.

Expected Time Complexity: O(N).
Expected Auxiliary Space: O(1).